requirements.txt
comtypes==1.4.8
dxcam==0.0.5
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.10.0
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pydirectinput_rgx==2.1.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.3
pytweening==1.2.0
pywin32==308
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.3.0
shapely==2.0.6
six==1.16.0
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
